Time: 15 ms (93.81%), Space: 14.1 MB (10.25%) - LeetHub

+++ 1669-minimum-cost-to-cut-a-stick/1669-minimum-cost-to-cut-a-stick.cpp
@@ -19,9 +19,9 @@ public:
         vector<vector<int>> dp(c+2, vector<int>(c+2, 0));
         for(int i = c-2; i >= 1; i--){
             for(int j = 1; j <= c-2; j++){
-                long long mini = INT_MAX;
+                int mini = INT_MAX;
                 for(int index = i; index <= j; index++){
-                    long long cost = cuts[j+1]-cuts[i-1] + dp[i][index-1] + dp[index+1][j];
+                    int cost = cuts[j+1]-cuts[i-1] + dp[i][index-1] + dp[index+1][j];
                     mini = min(cost, mini);
                 }
                 if(mini != INT_MAX){
